feat: 업체 조회 API 추가
-[] getCompany() 추가
-[] createCompany(), getCompany() 가 공통 응답 객체를 반환하도록 수정
