+++ charlee-web/src/main/java/com/efa73/charleeweb/company/domain/repository/CompanyRepository.java
@@ -1,6 +1,6 @@
-package com.efa73.charleeweb.company.repository;
+package com.efa73.charleeweb.company.domain.repository;
 
-import com.efa73.charleeweb.company.entity.Company;
+import com.efa73.charleeweb.company.domain.entity.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CompanyRepository extends JpaRepository<Company, Long> {
