+++ charlee-web/src/main/java/com/efa73/charleeweb/company/interfaces/dto/CompanyRequest.java
@@ -1,6 +1,6 @@
-package com.efa73.charleeweb.company.dto;
+package com.efa73.charleeweb.company.interfaces.dto;
 
-import com.efa73.charleeweb.company.entity.Company;
+import com.efa73.charleeweb.company.domain.entity.Company;
 
 public record CompanyRequest(String name, String siteLink) {
